[IMP] purchase: show all pickings in purchase order
When the multi step reception is activated on a warehouse, we do not see
all the reception steps, we only see the ones that are the first step of
reception.

To show all the reception related to the PO, we will show all the
receptions that have the same procurement group than the PO.

We will hide the button 'Receive Products' once all the picking of first
step are done or cancel.