[IMP] tests: Lazily evaluate Odoo port
When importing something from common.py, the Odoo http port was
taken from the command line parameters. When importing too early,
the port was set to the default one. For example, when importing in
module.py before the command line parameters are evaluated.
If that happens, it makes the runbot red.
This commit makes a lazy evaluation of the port when needed.
Make the runbot green again.