[MOV] hr_holidays: split data to separate mail data from other data
It is simpler to have data split by main model or application like we
already do for models and views. It allows to easily have an overview
of data a module holds. Moreover as we will add mail related data having
it all in a single file and not lost between other data helps finding
and working with it.

This commit only moves code. No functional change should occur.