[IMP] core: menu tip ux/ui improvement
We would like to make the "no item found" screens more appealing.

Before this commit, it shows a small help tip in the top-left
corner of the screen, just below the "Create" button.

With this commit, these help tips have been replaced by onboarding
screens, which consist of a picture and some text below, both of which
are horizontally centered.

The texts have been slightly changed, so that they are shorter and clearer.
Also, they include a link to create an item, of which their behavior is
exactly similar to the "create" buttons.

For the moment, only the following apps have been considered:
    - Project
    - Sales
    - Inventory
    - MRP
    - Leaves
    - Invoicing
    - eCommerce
    - Attendances
    - Expenses
    - Employees
    - Contacts
    - Mass Mailing
    - Survey
    - Link Tracker
    - CRM
    - Maintenance
    - Lunch
    - Fleet
    - IM Livechat
    - Events
    - POS
    - Purchases
    - PLM
    - Helpdesk
    - Subscription
    - Accounting