[IMP] test_performance: improve base class and add mail performance tests
In test_performance base class holding tools methods is separated from
the class holding tests. It allows to reuse the base class without having
to depend on existing tests.

This commit updates test_mail to use the class with tests. This means that
existing base performance test are recomputed in test_mail. It allows to
test that mail did not add unnecessary queries to base performance test.

This commit also add tools method to halt and resume query count. It allows
to perform asserts or update some data without counting the queries
necessary to do it.

New tests are added in test_mail to have tests for message_post and complex
mail thread features.