This commit emphasize the fact that expense policy is wrongly use
Indeed it is used to determine the computation of a
SO line (with expense product): it only depends on the
analytic sum of unit_amount. This make sense when a SO
line is created from hr.expense or vendor bill (we don't
the delivered quantity to be maunally updateable to force
people to use mentionned mecanisms).
As consequence, we can't sell properly a expensable product,
since the delivered quantity change with its link analytic lines
even if the product is 'stockable' or 'service-timesheet'.

--> a flag on sale.order.line saying if the line comes from an
expense/vendor bills can avoid duplicating all products (expensable
version and sellable version). This solution impact the line of this
commit, I swear.