[IMP] sale,hr_expense: reinvoice policy and 'can be expensed'
Impacted modules: sale, hr_expense, sale_stock and
sale_expense.

The reinvoice policy has nothing to do with the
fact a product can be expense or not by employee,
because of the vendor bills.
You can only choose reinvoice policy of your product
when hr_expense is installed, which is not very intuitive
for users. So this commit make reinvoice policy field
always visible when sale is installed. 'can_be_expensed'
field determines if the product can be used in HR expense.

However, since reinvoice policy does not make much sense
for stockable products, we hide this and reset it to 'no'
in case of consumable and stockable products when sale_stock
module is installed, as we don't handle stock for those
expensed products.