[FIX] barcodes: use patch to prevent interactions between tests
There is a subtle issue happening in tests when we override a widget
property for the duration of the test and tries to restore the initial
state at the end by rewriting on the property.  It actually changes the
widget, by having a function bound to the sub widget prototype instead
of its ancestor.  This in turns can cause weird interactions with the
class system.

We introduced a proper way to temporarily patch a class/widget by adding
a method in testUtils to do exactly that.  This commit simply uses that
method to prevent unwanted interactions in barcode tests