[IMP] sale: vendor bills and expense flow
Vendor bills and expense use the same mecanism to be reinvoiced: the
AA creation may trigger a SOL creation, based on the reinvoice policy
and AA of SO from invoice lines or hr expense.
We only increment expense SOL (is_expense=True), but mostly a new line
will be created on the SO.

Here are all the use cases:

Re-invoiced at cost :
1. No expense SOL, ordered quantity: add a new line. Ordered quantity =
0,
Delivered quantity = expense quantity
2. No expense SOL, delivered quantity: add a new line. Ordered quantity
= 0,
 quantity = expense quantity
3. Expense SOL with a re-invoiced at cost: The use case doesn't make
sense
because the expenses can have several cost, so it shouldn't be planned
on
the SO. In any case, add a new line with Ordered quantity = 0 and
Delivered
quantity = expense quantity

Re-invoiced at sales price :
1. No expense SOL, ordered quantity: add a new line. Ordered quantity =
0,
Delivered quantity = expense quantity
2. No expense SOL, delivered quantity: add a new line. Ordered quantity
= 0,
Delivered quantity = expense quantity
3. SO line with the expense product, ordered quantity : the delivered
quantity
should be incremented when an expense with the matching product (linked
to
that SO, no matter the description of the expense) is submitted.