[IMP] hr_timesheet: ability to track cost with selling time
The cost of an employee should be in hr_timesheet for a long time, to
allow user to track time on project/task: hours spent and their cost.
This commit simply move fields and related views. It also move the
 method from sale_timesheet to hr_timesheet
since this method is responsible to set the amount (cost) on AAL.