[IMP] website, *: allow parallax on all snippets
* web_editor

Before this commit, the parallax effect was specific to a given snippet
(the parallax one). This in fact did not make much sense as this is an
effect which can be added on all sections which have a background image.
The parallax option is here extended to be present on all snippets and
allows to be enabled/disabled.

Note: the parallax snippet itself is kept for now as it has an
additional feature: being able to contain multiple snippets and have the
background effect to apply accross all of them.

Note2: this commit breaks the compatibility of old parallax snippets
which were added in client websites. As this is a non critical feature
and that the parallax effect can simply be reenabled with the editor,
this should not be a big problem.

See API changes that made the implementation of this feature possible:
...
...