[REM] mail: remove 'message_last_post' mechanism
Thread mixin class adds a message_last_post field. This field may be updated when a
comment is posted, depending on a context key. Since a while this context key is
not used leading to the field not being updated. Moreover this field is used only
in a filter in project and do not have any other use.

This commit remove the field and the context key. This feature can be achieved
in custom easily if necessary.