[FIX] web: bin_size set for rpc read/search_read on lists
Data of binary fields were downloaded by default when accessing
a list view with binary fields. This behaviour can be disabled
with {bin_size: True} in the context.

Now, this contextual item is added and set to true whenever
calling read or search_read on list views.

We could have let some fields making changes in the context,
such as by enforcing {bin_size: True} whenever we load at least
one binary field.

We decided to not do so for the following reasons:
    - only use-case so far is with binary fields and bin_size,
    - more complexity on the fields and the views,
    - non-negligible drop of performance with rpc calls on lists.

We might change our mind and add this feature in the future.